pyproject.toml
[project]
name = "rarity-api"
version = "0.1.0"
description = ""
authors = [
    {name = "artem",email = "artem@deffun.io"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "stripe (>=12.0.0,<13.0.0)",
]

[tool.poetry]
packages = [{include = "rarity_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
